Refactor dora meta manager
### What changes are proposed in this pull request?

1. Move metastore into dora meta manager
2. Handles listing cache invalidation
3. Add unit tests
4. Introduced caffeinine cache

### Why are the changes needed?

Dora metadata management has a lot of consistency issue between metadata/listing/page cahces. To address these issues, the first step is to introduce a centralized management place and put these logic together.

### Does this PR introduce any user facing changes?

N/A
			pr-link: Alluxio/alluxio#17710

+++ dora/core/server/worker/src/main/java/alluxio/worker/dora/PagedDoraWorker.java
@@ -31,7 +31,6 @@ import alluxio.exception.AccessControlException;
 import alluxio.exception.FileAlreadyExistsException;
 import alluxio.exception.runtime.AlluxioRuntimeException;
 import alluxio.exception.status.NotFoundException;
-import alluxio.file.FileId;
 import alluxio.grpc.Command;
 import alluxio.grpc.CommandType;
 import alluxio.grpc.CompleteFilePOptions;
@@ -72,7 +71,6 @@ import alluxio.underfs.UfsStatus;
 import alluxio.underfs.UnderFileSystem;
 import alluxio.underfs.UnderFileSystemConfiguration;
 import alluxio.underfs.options.CreateOptions;
-import alluxio.underfs.options.ListOptions;
 import alluxio.util.CommonUtils;
 import alluxio.util.ModeUtils;
 import alluxio.util.executor.ExecutorServiceFactories;
@@ -89,8 +87,6 @@ import alluxio.worker.task.DeleteHandler;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
-import com.google.common.cache.Cache;
-import com.google.common.cache.CacheBuilder;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.Closer;
@@ -103,7 +99,6 @@ import org.slf4j.LoggerFactory;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -111,6 +106,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicReference;
+import javax.annotation.Nullable;
 import javax.inject.Named;
 
 /**
@@ -133,20 +129,7 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
   private final String mRootUFS;
   private FileSystemContext mFsContext;
 
-  private static class ListStatusResult {
-    public long mTimeStamp;
-    public UfsStatus[] mUfsStatuses;
-
-    ListStatusResult(long timeStamp, UfsStatus[] ufsStatuses) {
-      mTimeStamp = timeStamp;
-      mUfsStatuses = ufsStatuses;
-    }
-  }
-
-  private final Cache<String, ListStatusResult> mListStatusCache;
   private WorkerNetAddress mAddress;
-
-  private DoraMetaStore mMetaStore;
 
   private final UnderFileSystem mUfs;
 
@@ -174,20 +157,12 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
     mUfs = UnderFileSystem.Factory.create(
         mRootUFS,
         UnderFileSystemConfiguration.defaults(Configuration.global()));
-    mListStatusCache = CacheBuilder.newBuilder()
-        .maximumSize(Configuration.getInt(PropertyKey.DORA_UFS_LIST_STATUS_CACHE_NR_DIRS))
-        .expireAfterWrite(Configuration.getDuration(PropertyKey.DORA_UFS_LIST_STATUS_CACHE_TTL))
-        .build();
 
     mPageSize = Configuration.global().getBytes(PropertyKey.WORKER_PAGE_STORE_PAGE_SIZE);
     mBlockMasterClientPool = new BlockMasterClientPool();
-
-    String dbDir = Configuration.getString(PropertyKey.DORA_WORKER_METASTORE_ROCKSDB_DIR);
-    Duration duration = Configuration.getDuration(PropertyKey.DORA_WORKER_METASTORE_ROCKSDB_TTL);
-    long ttl = (duration.isNegative() || duration.isZero()) ? -1 : duration.getSeconds();
-    mMetaStore = new RocksDBDoraMetaStore(dbDir, ttl);
     mCacheManager = cacheManager;
-    mMetaManager = new DoraMetaManager(this, mMetaStore, mCacheManager, mUfs);
+    mMetaManager = mResourceCloser.register(
+        new DoraMetaManager(this, mCacheManager, mUfs));
     mOpenFileHandleContainer = new DoraOpenFileHandleContainer();
   }
 
@@ -260,9 +235,6 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
 
   @Override
   public void close() throws IOException {
-    if (mMetaStore != null) {
-      mMetaStore.close();
-    }
     try (AutoCloseable ignoredCloser = mResourceCloser;
          AutoCloseable ignoredCacheManager = mCacheManager
     ) {
@@ -280,6 +252,7 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
   }
 
   @Override
+  @Nullable
   public UfsStatus[] listStatus(String path, ListStatusPOptions options)
       throws IOException, AccessControlException {
     final long syncIntervalMs = options.hasCommonOptions()
@@ -287,65 +260,19 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
         ? options.getCommonOptions().getSyncIntervalMs() : -1) :
         -1;
 
-    final boolean skipCache = options.hasRecursive() && options.getRecursive();
-    final UfsStatus[] cachedStatuses;
-    final ListStatusResult resultFromCache = mListStatusCache.getIfPresent(path);
-    if (resultFromCache == null) {
-      cachedStatuses = null;
-    } else if (skipCache) {
-      // Only use the cached result when its not recursive listing
-      cachedStatuses = null;
-    } else {
-      // Metadata is cached. Check if it is expired.
-      if (syncIntervalMs >= 0
-          && System.nanoTime() - resultFromCache.mTimeStamp > syncIntervalMs * Constants.MS_NANO) {
-        // The metadata is expired. Remove it from in-memory cache.
-        mListStatusCache.invalidate(path);
-        cachedStatuses = null;
-      } else {
-        // Cache is still valid. Use cached statuses.
-        cachedStatuses = resultFromCache.mUfsStatuses;
-      }
-    }
-    if (cachedStatuses != null) {
-      return cachedStatuses;
-    }
+    boolean isRecursive = options.getRecursive();
+    final Optional<ListStatusResult> resultFromCache = mMetaManager.listCached(path, isRecursive);
 
-    // Not found in cache. Query the Under File System.
-    ListOptions ufsListOptions = ListOptions.defaults().setRecursive(
-        options.hasRecursive() ? options.getRecursive() : false);
-    UfsStatus[] freshStatusesFromUfs = mUfs.listStatus(path, ufsListOptions);
-
-    if (freshStatusesFromUfs == null) {
-      // If empty, the request path might be a regular file/object. Let's retry getStatus().
-      try {
-        UfsStatus status = mUfs.getStatus(path);
-        // Success. Create an array with only one element.
-        status.setName(""); // listStatus() expects relative name to the @path.
-        freshStatusesFromUfs = new UfsStatus[1];
-        freshStatusesFromUfs[0] = status;
-      } catch (FileNotFoundException e) {
-        // Do nothing.
-        // The freshStatusesFromUfs is still null to indicate empty listStatus() result.
-      }
+    if (resultFromCache.isPresent()
+        && (syncIntervalMs < 0
+        || System.nanoTime() - resultFromCache.get().mTimeStamp
+        <= syncIntervalMs * Constants.MS_NANO)) {
+      return resultFromCache.get().mUfsStatuses;
     }
-
-    // Add this into cache. Return value of listStatus() might be null if not found.
-    if (freshStatusesFromUfs != null && !skipCache) {
-      ListStatusResult newResult = new ListStatusResult(System.nanoTime(), freshStatusesFromUfs);
-      mListStatusCache.put(path, newResult);
-    }
-    return freshStatusesFromUfs;
-  }
-
-  /**
-   * Invalidate the given cached File by deleting it from local cache.
-   *
-   * @param path the full path of this file
-   */
-  private void invalidateCachedFile(String path) {
-    FileId file = FileId.of(new AlluxioURI(path).hash());
-    mCacheManager.deleteFile(file.toString());
+    mMetaManager.invalidateListingCache(path);
+    Optional<UfsStatus[]> ufsStatuses =
+        mMetaManager.listFromUfsThenCache(path, isRecursive);
+    return ufsStatuses.orElse(null);
   }
 
   @Override
@@ -696,8 +623,7 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
             new FileAlreadyExistsException("File already exists but no overwrite flag"));
       } else if (overWrite) {
         // client is going to overwrite this file. We need to invalidate the cached meta and data.
-        invalidateFileMeta(path);
-        invalidateCachedFile(path);
+        mMetaManager.removeFromMetaStore(path);
       }
 
       // Open UFS OutputStream and use it in write operation.
@@ -727,20 +653,6 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
     return handle;
   }
 
-  private void invalidateFileMeta(String path) {
-    // The simplest way of updating metadata is invalidating cache in worker.
-    // Next time, worker will get fresh metadata from ufs.
-    AlluxioURI fullPathUri = new AlluxioURI(path);
-    AlluxioURI parentDir;
-    if (fullPathUri.isRoot()) {
-      parentDir = fullPathUri;
-    } else {
-      parentDir = fullPathUri.getParent();
-    }
-    mListStatusCache.invalidate(parentDir.toString()); // invalidate dir cache
-    mMetaStore.removeDoraMeta(path);                   // invalidate in-Rocks cache
-  }
-
   @Override
   public void completeFile(String path, CompleteFilePOptions options, String uuid)
       throws IOException, AccessControlException {
@@ -749,6 +661,7 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
       mOpenFileHandleContainer.remove(path);
       handle.close();
       Optional<DoraMeta.FileStatus> status = mMetaManager.loadFromUfs(path);
+      mMetaManager.invalidateListingCacheOfParent(path);
       if (status.isEmpty()) {
         throw new FileNotFoundException("Cannot retrieve file metadata of "
             + path + " when completing the file");
@@ -760,16 +673,16 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
   public void delete(String path, DeletePOptions options) throws IOException,
       AccessControlException {
     try {
-      invalidateFileMeta(path);
-      invalidateCachedFile(path);
+      mMetaManager.removeFromMetaStore(path);
 
       // TODO(hua) Close the open file handle?
-
-      UfsStatus status = mUfs.getStatus(path);
-      if (status.isFile()) {
-        mUfs.deleteFile(path);
-      } else {
-        mUfs.deleteDirectory(path);
+      if (!options.getAlluxioOnly()) {
+        UfsStatus status = mUfs.getStatus(path);
+        if (status.isFile()) {
+          mUfs.deleteFile(path);
+        } else {
+          mUfs.deleteDirectory(path);
+        }
       }
     } catch (IOException e) {
       throw new RuntimeException(e);
@@ -780,17 +693,15 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
   public void rename(String src, String dst, RenamePOptions options)
       throws IOException, AccessControlException {
     try {
-      invalidateFileMeta(src);
-      invalidateCachedFile(src);
-      invalidateFileMeta(dst);
-      invalidateCachedFile(dst);
-
       UfsStatus status = mUfs.getStatus(src);
       if (status.isFile()) {
         mUfs.renameFile(src, dst);
       } else {
         mUfs.renameDirectory(src, dst);
       }
+      mMetaManager.removeFromMetaStore(src);
+      mMetaManager.loadFromUfs(dst);
+      mMetaManager.invalidateListingCacheOfParent(dst);
     } catch (IOException e) {
       throw new RuntimeException(e);
     }
@@ -800,10 +711,9 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
   public void createDirectory(String path, CreateDirectoryPOptions options)
       throws IOException, AccessControlException {
     try {
-      invalidateFileMeta(path);
-      invalidateCachedFile(path);
-
       boolean success = mUfs.mkdirs(path);
+      mMetaManager.loadFromUfs(path);
+      mMetaManager.invalidateListingCacheOfParent(path);
       if (!success) {
         throw new RuntimeException(
             new FileAlreadyExistsException(String.format("%s already exists", path)));
@@ -851,6 +761,7 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
       mUfs.setOwner(path, null, options.getGroup());
     }
     mMetaManager.loadFromUfs(path);
+    mMetaManager.invalidateListingCacheOfParent(path);
   }
 
   @Override
@@ -894,6 +805,11 @@ public class PagedDoraWorker extends AbstractWorker implements DoraWorker {
   @VisibleForTesting
   UnderFileSystem getUfs() {
     return mUfs;
+  }
+
+  @VisibleForTesting
+  DoraMetaManager getMetaManager() {
+    return mMetaManager;
   }
 
   protected void checkCopyPermission(String srcPath, String dstPath)
